refactor: ♻️ move plugin-context.ts to types
